+++ doc/source/conf.py
@@ -26,7 +26,7 @@ author = 'Bogdan Opanchuk'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '0.0.2'
+release = '0.0.3'
 
 
 # -- General configuration ---------------------------------------------------