+++ doc/source/history.rst
@@ -3,8 +3,8 @@ Version history
 ---------------
 
 
-0.0.3 (current development version)
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+0.0.3 (19 Jul 2019)
+~~~~~~~~~~~~~~~~~~~
 
 * ADDED: ``LweSampleArray.copy()`` for cloning a ciphertext.